feat: add initial implementation of the bank
A ledger is read from a csv which contains all the transactions.
The bank can then iterate through the transactions and make up "virtual" account balances/states
as it goes through.

a number of test cases have added, though we're still missing a few frozen/locked account test,
precision and tests with large datasets but at this point I can't have a play with that.